Added IMath library to replace & complete the bigint implementation. Added a mention in the README and the copyright info in COPYRIGHT.
Rationale: IMath is actively maintained and also contains a rational implementation that will come handy. It is distributed under the MIT license, just as Lua and klisp. The main necessary modification is that the klisp allocator should be used instead of malloc. This modification is easy but tedious, nonetheless the benefits of using an already implemented and tested library are huge.